requirements.txt
pydantic==1.10.6
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.5post1
beautifulsoup4~=4.11.1
# playwright==1.30.0
